strands-demo/demo_agentic_search/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
PyPDF2>=3.0.0
